+++ src/main/java/com/drmangotea/tfmg/content/machinery/metallurgy/blast_furnace/BlastFurnaceOutputBlockEntity.java
@@ -100,6 +100,12 @@ public class BlastFurnaceOutputBlockEntity extends SmartBlockEntity implements I
         );
     }
 
+    public void addFuel(int amount) {
+        this.fuel += amount;
+        setChanged();
+        sendData();
+    }
+
     @Override
     public void addBehaviours(List<BlockEntityBehaviour> behaviours) {
     }